feat: 更新continew-admin-ui版本至v3.6

+++ app-service/vertex-framework/vertex-framework-common/src/main/java/com/onezol/vertex/framework/common/constant/DefaultPage.java
@@ -8,20 +8,20 @@ public final class DefaultPage {
     /**
      * 默认页码
      */
-    public static int DEFAULT_PAGE_NUMBER = 1;
+    public static long DEFAULT_PAGE_NUMBER = 1;
     /**
      * 默认页大小
      */
-    public static int DEFAULT_PAGE_SIZE = 10;
+    public static long DEFAULT_PAGE_SIZE = 10;
 
     /**
      * 最大页大小
      */
-    public static int MAX_PAGE_SIZE = 100;
+    public static long MAX_PAGE_SIZE = 100;
 
     /**
      * 每页条数 - 不分页(常用于导出接口)
      */
-    public static int PAGE_SIZE_NONE = -1;
+    public static long PAGE_SIZE_NONE = -1;
 
 }
